refactor: most of the domain-related types
