+++ packages/encodable/src/parsers/domain/parseDiscreteDomain.ts
@@ -1,10 +1,12 @@
-import { ChannelInput } from '../../types/Channel';
+import { DiscreteInput } from '../../types/ScaleConfig';
 
 /**
  * Discrete domains are converted into string[]
  * when using D3 scales
  * @param domain
  */
-export default function parseDiscreteDomain<T extends ChannelInput>(domain: T[]) {
+export default function parseDiscreteDomain<T extends DiscreteInput | undefined | null>(
+  domain: T[],
+) {
   return domain.map(d => `${d}`);
 }
