+++ packages/encodable/src/parsers/scale/applyInterpolate.ts
@@ -1,48 +1,8 @@
-import {
-  interpolateRgb,
-  interpolateLab,
-  interpolateHcl,
-  interpolateHclLong,
-  interpolateHsl,
-  interpolateHslLong,
-  interpolateCubehelix,
-  interpolateCubehelixLong,
-} from 'd3-interpolate';
 import { InterpolatorFactory } from 'd3-scale';
 import { Value } from '../../types/VegaLite';
 import { D3Scale } from '../../types/Scale';
-import { ScaleInterpolateParams, ScaleInterpolate } from '../../types/scale/ScaleInterpolate';
 import { ScaleConfig } from '../../types/ScaleConfig';
-
-const interpolatorMap = {
-  lab: interpolateLab,
-  hcl: interpolateHcl,
-  'hcl-long': interpolateHclLong,
-  hsl: interpolateHsl,
-  'hsl-long': interpolateHslLong,
-  cubehelix: interpolateCubehelix,
-  'cubehelix-long': interpolateCubehelixLong,
-  rgb: interpolateRgb,
-} as const;
-
-function getInterpolator(interpolate: ScaleInterpolate | ScaleInterpolateParams) {
-  switch (interpolate) {
-    case 'lab':
-    case 'hcl':
-    case 'hcl-long':
-    case 'hsl':
-    case 'hsl-long':
-    case 'cubehelix':
-    case 'cubehelix-long':
-    case 'rgb':
-      return interpolatorMap[interpolate];
-    default:
-  }
-
-  const { type, gamma } = interpolate;
-  const interpolator = interpolatorMap[type];
-  return typeof gamma === 'undefined' ? interpolator : interpolator.gamma(gamma);
-}
+import createInterpolator from './createInterpolator';
 
 export default function applyInterpolate<Output extends Value>(
   config: ScaleConfig<Output>,
@@ -54,7 +14,7 @@ export default function applyInterpolate<Output extends Value>(
     'interpolate' in scale
   ) {
     scale.interpolate(
-      (getInterpolator(config.interpolate) as unknown) as InterpolatorFactory<Output, Output>,
+      (createInterpolator(config.interpolate) as unknown) as InterpolatorFactory<Output, Output>,
     );
   }
 }
