+++ packages/encodable/src/parsers/scale/createInterpolator.ts
@@ -1,0 +1,41 @@
+import {
+  interpolateRgb,
+  interpolateLab,
+  interpolateHcl,
+  interpolateHclLong,
+  interpolateHsl,
+  interpolateHslLong,
+  interpolateCubehelix,
+  interpolateCubehelixLong,
+} from 'd3-interpolate';
+import { ScaleInterpolateParams, ScaleInterpolate } from '../../types/scale/ScaleInterpolate';
+
+const interpolatorMap = {
+  lab: interpolateLab,
+  hcl: interpolateHcl,
+  'hcl-long': interpolateHclLong,
+  hsl: interpolateHsl,
+  'hsl-long': interpolateHslLong,
+  cubehelix: interpolateCubehelix,
+  'cubehelix-long': interpolateCubehelixLong,
+  rgb: interpolateRgb,
+} as const;
+
+export default function createInterpolator(interpolate: ScaleInterpolate | ScaleInterpolateParams) {
+  switch (interpolate) {
+    case 'lab':
+    case 'hcl':
+    case 'hcl-long':
+    case 'hsl':
+    case 'hsl-long':
+    case 'cubehelix':
+    case 'cubehelix-long':
+    case 'rgb':
+      return interpolatorMap[interpolate];
+    default:
+  }
+
+  const { type, gamma } = interpolate;
+  const interpolator = interpolatorMap[type];
+  return typeof gamma === 'undefined' ? interpolator : interpolator.gamma(gamma);
+}
